Updated styling for button
The background of the button is a image now and I removed focus and replaced it with active, now when you click it there is a slight animation. If this is not what you had in mind please tell me what you had in mind and I would love to implement that to. Also if you want to see the button in action check this this demo out: "https://codepen.io/guildshark/pen/mddyeqm?editors=1100".